+++ src/helpers/analytics/trackGAEvent.ts
@@ -1,15 +1,8 @@
 import ReactGA from "react-ga";
 
-import { Environment } from "./environment/Environment/Environment";
+import { Environment } from "../environment/Environment/Environment";
 
 const GA_API_KEY = Environment.getGoogleAnalyticsApiKey();
-
-declare global {
-  interface Window {
-    analytics: any; // Segment.js
-    gtag: any; // Google Tag Manager
-  }
-}
 
 export const trackGAEvent = (event: ReactGA.EventArgs) => {
   try {
@@ -17,7 +10,7 @@ export const trackGAEvent = (event: ReactGA.EventArgs) => {
     if (GA_API_KEY && ReactGA) {
       ReactGA.event(event);
     }
-  } catch (e) {
-    console.error("trackGAEvent", e);
+  } catch (error) {
+    console.error("trackGAEvent", error);
   }
 };
